Support Iterable type in algorithm input signatures and support extract_subgraph for NetworkX
This commit adds support for extract_subgraph for NetworkX.

Since extract_subgraph takes 2 inputs, a graph and an Iterable of nodes, we need to support Iterable[...] in the abstract and concrete input signatures.

This commit accomplishes that.

+++ metagraph/tests/algorithms/test_betweenness_centrality.py
@@ -43,7 +43,7 @@ v        v /        v
         6: 11.0,
     }
     expected_answer = dpr.wrapper.Nodes.PythonNodes(
-        expected_answer_unwrapped, dtype="int"
+        expected_answer_unwrapped, dtype="float"
     )
     MultiVerify(
         dpr,
